packages/ai/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
docopt==0.6.2
filelock==3.13.1
Flask==3.0.0
fsspec==2023.12.2
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
minio==7.2.0
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai-whisper==20231117
pika==1.3.2
pipreqs==0.4.13
pycparser==2.21
pycryptodome==3.19.0
pyparsing==3.1.1
pyrabbit==1.1.0
regex==2023.10.3
requests==2.31.0
sympy==1.12
tiktoken==0.5.2
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
yarg==0.1.9
zipp==3.17.0
